Move editor associations to override service
