+++ src/vs/workbench/contrib/extensions/browser/extensions.web.contribution.ts
@@ -6,9 +6,10 @@
 import { localize } from 'vs/nls';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
-import { EditorDescriptor, IEditorRegistry, Extensions as EditorExtensions } from 'vs/workbench/browser/editor';
+import { EditorDescriptor, IEditorRegistry } from 'vs/workbench/browser/editor';
 import { RuntimeExtensionsEditor } from 'vs/workbench/contrib/extensions/browser/browserRuntimeExtensionsEditor';
 import { RuntimeExtensionsInput } from 'vs/workbench/contrib/extensions/common/runtimeExtensionsInput';
+import { EditorExtensions } from 'vs/workbench/common/editor';
 
 // Running Extensions
 Registry.as<IEditorRegistry>(EditorExtensions.Editors).registerEditor(
